share with team

+++ main/views/projects.py
@@ -5,7 +5,7 @@ from django.utils.text import slugify
 from django.shortcuts import render
 from django.http import JsonResponse
 
-from main.models import Project, ProjectUser, Team, TeamUser
+from main.models import Project, ProjectUser, Team, TeamUser, ProjectTeam
 from main.utils import get_project, put_project_file
 
 import shutil
@@ -14,12 +14,22 @@ import os
 
 @login_required
 def list_projects(request):
+    template = 'projects/list.html'
+
+    if request.POST.get('action') == 'delete':
+        project = get_project(request.user.username, request.POST['name'])
+
+        if project:
+            shutil.rmtree(project.get_path())
+            project.delete()
+            template = '_partial/_list_projects.html'
+
     context = {
         'projects': Project.objects.filter(user=request.user),
         'shared_with_me': False,
         'page_title': 'My Projects'
     }
-    return render(request, 'projects/list.html', context)
+    return render(request, template, context)
 
 
 @login_required
@@ -50,53 +60,74 @@ def edit_project(request, project_name):
 @login_required
 def share_project(request, project_name):
     project = get_project(request.user.username, project_name)
-    template = 'projects/share.html'
 
-    if request.method == "POST":
-        if request.POST.get('action') == 'generate_view_key':
-            project.view_key = get_random_string(length=16)
-            project.save()
-            context = {
-                'status': 'ok',
-                'view_key': project.view_key
-            }
-            return JsonResponse(context)
+    if request.POST.get('action') == 'generate_view_key':
+        project.view_key = get_random_string(length=16)
+        project.save()
+        context = {
+            'status': 'ok',
+            'view_key': project.view_key
+        }
+        return JsonResponse(context)
 
-        if request.POST.get('action') == 'share_with_email':
-            template = '_partial/_shared_email_list.html'
-            email = request.POST.get('email').strip()
-            already_shared = ProjectUser.objects.filter(project=project, email__iexact=email)
-            if not already_shared and email:
-                ProjectUser(project=project, email=email).save()
+    shared_emails = ProjectUser.objects.filter(project=project)
+    shared_teams = ProjectTeam.objects.filter(project=project)
 
-        if request.POST.get('action') == 'delete_email':
-            template = '_partial/_shared_email_list.html'
-            email = request.POST.get('email')
-            sharing = ProjectUser.objects.filter(project=project, email__iexact=email)
-            if sharing:
-                context = {'status': 'ok'}
+    if request.POST.get('action') == 'share_with_email':
+        email = request.POST.get('email').strip()
+        already_shared = ProjectUser.objects.filter(project=project, email__iexact=email)
+
+        if not already_shared and email:
+            ProjectUser(project=project, email=email).save()
+
+        return render(request, '_partial/_shared_email_list.html', {'shared_emails': shared_emails})
+
+    if request.POST.get('action') == 'delete_email':
+        email = request.POST.get('email')
+        sharing = ProjectUser.objects.filter(project=project, email__iexact=email)
+
+        if sharing:
+            sharing.delete()
+
+        return render(request, '_partial/_shared_email_list.html', {'shared_emails': shared_emails})
+
+    if request.POST.get('action') == 'share_with_team':
+        team_id = request.POST.get('team_id')
+
+        if team_id:
+            team = Team.objects.get(id=team_id)
+
+            is_member = TeamUser.objects.filter(user=request.user, team=team)
+            already_shared = ProjectTeam.objects.filter(project=project, team=team)
+
+            if not already_shared and is_member:
+                ProjectTeam(project=project, team=team).save()
+
+        return render(request, '_partial/_shared_team_list.html', {'shared_teams': shared_teams})
+
+    if request.POST.get('action') == 'delete_team':
+        team_id = request.POST.get('team_id')
+
+        if team_id:
+            team = Team.objects.get(id=team_id)
+
+            is_member = TeamUser.objects.filter(user=request.user, team=team)
+            sharing = ProjectTeam.objects.filter(project=project, team=team)
+
+            if sharing and is_member:
                 sharing.delete()
+
+        return render(request, '_partial/_shared_team_list.html', {'shared_teams': shared_teams})
+
+    teams = Team.objects.filter(teamuser__user=request.user)
 
     context = {
         'project': project,
-        'shared_emails': ProjectUser.objects.filter(project=project),
-        'teams': Team.objects.filter(teamuser__user__id=request.user.id)
+        'shared_emails': shared_emails,
+        'shared_teams': shared_teams,
+        'teams': teams
     }
-    return render(request, template, context)
-
-
-@login_required
-@require_POST
-def delete_project(request):
-    project = get_project(request.user.username, request.POST['name'])
-
-    if project:
-        shutil.rmtree(project.get_path())
-        project.delete()
-        return JsonResponse({'status': 'ok'})
-
-    return JsonResponse({'status': 'error'})
-
+    return render(request, 'projects/share.html', context)
 
 @login_required
 def new_project(request):
